Upgraded to ABP 9.1

+++ aspnet-core/src/EventCloud.Application/Events/EventAppService.cs
@@ -61,7 +61,7 @@ namespace EventCloud.Events
 
         public async Task CreateAsync(CreateEventInput input)
         {
-            var tenantId = (int)AbpSession.TenantId.Value;
+            var tenantId = AbpSession.TenantId.Value;
             var @event = Event.Create(tenantId, input.Title, input.Date, input.Description, input.MaxRegistrationCount);
             await _eventManager.CreateAsync(@event);
         }
